Linear system file read refactor (#3)
This branch makes signficant chages to how IJ and matrix market files are read in. First, the size of the linear system (per rank) is determined from the file headers. Then, the data for the matrices and vectors is read into a large buffer in CPU memory first in COO format. Then, the data is passed into the Hypre SetValues2 API (matrix or vector) as either a host memory pointer (CPU) or a UVM pointer (CUDA, GPU). This enables matrix assembly, and precon setup to run on GPUs.